requirements.txt
google-api-python-client==2.149.0
langdetect==1.0.9
googletrans==4.0.0-rc1
textblob==0.18.0.post0
nltk==3.9.1
flask==3.0.3
pandas==2.2.3
pyyaml==6.0.2